requirements.txt
Flask
Flask-Script
Flask-Login
Flask-SQLAlchemy
Flask-WTF
python-ldap
SQLAlchemy
MySQL-python
WTForms
TestLink-API-Python-client
redis